requirements.txt
aiogram>=3.4.1
ffmpeg-python>=0.2.0
python-dotenv>=1.0.1
